Updates
Added in mobs for new bashing zones (thanks Mazzsys).
Updated most 'touch shield' in auto-basher to make sure you don't already have it deffed before trying to do it again to stop so much spam.
Put in experimental clot and shield into basher for Zealot.
Updated new target auto code so that if you're not calling, it'll try to still work through the target list. Just in case callers die. Experimental.
Added in an 'all' for Vortex bashing. Just for me,  but you could update for yourself.
Added in 'tlist' to trigger to call target list.