Rihrin Stuff
More bashing adjustments for Zealot for Rihrin.

To be clear here: my changes have been to add in ways to track punisher npcs so you don't accidentally shield while bashing them. But, all my code right now requires a wand of reflection and a lyre. So, best not to update if you play Zealot. And, none of the code will affect you otherwise unless you go code it yourself.