Revenant Fixes
I still need to figure out if limbs are tracking well enough, but I tested all aliases. They are good to go for group combat, but need thorough testing in those experiences.

Also fixed an error with discern that was not tracking limb affs because of the limb update. And, forced Telvi to track - not sure why it won't how it is right now unless I use 'start of string'. More to come.

Summary: Revenant is playable.