Updates Again
After reviewing and having a fight with Benedicto, I realize that you all might not cure well from Warden's. So, I'm adding in the 'single function' tracking for classes in a hope that it'll give you all some headway in that regard.

It will need testing and there will be some heavy overlap for some classes where I'll have to delete out triggers. This will take a long time.