Echo Updates
For being killed, I'm adding in an echo to confirm that the found_set and snd.class_hitting are matching. This will help me determine what's causing errors with the firstclass reset. Please let me know if it ever does not match. And, if it does and you still get the reset issues, let me know.