Firstaidassist Adjustments
I was using the wrong tracking for how many people were hitting. Using that, and a few other adjustments, to try to make it so the system doesn't switch curesets if you have too many people hitting you. If it works, I'll build out a 'general group' set of some sort.