Earthcaller
This is not complete, but the foundation to be able to use Earthcaller is started. Here's what's done:
-Aliases for combat.
-Defenses.
-Bashing.
-Sch alias update for that class.
-All attacks updated.
-A trigger or two updated to include what I know of Earthcallers, but will need more.
-Firstaidassist updates to include Earthcaller.

Unrelated fix - my system is breaking sometimes, weirdly, so removed something from the primary prompt to see if it fixes it.  Also updated all #people_hitting_me to a count version that is used in normal Sunder.